ctap: reject duplicate CBOR map keys above 15
zf_ctap_mark_seen_key() tracked seen keys in a uint16_t and returned true
("not seen yet") for every key >= 16, so the strict duplicate-key rejection
that makeCredential, getAssertion and authenticatorClientPIN rely on silently
did not apply to the upper key space: an untrusted host could repeat a high
key as often as it liked. Widen the mask and the guard to 32 bits, which
covers every key those commands define, and widen the three callers' locals
to match. Keys >= 32 bind no request state and stay accepted.

+++ src/ctap/parse/get_assertion.c
@@ -69,7 +69,7 @@ uint8_t zf_ctap_parse_get_assertion(const uint8_t *data, size_t size,
                                     ZfGetAssertionRequest *request) {
     ZfCborCursor cursor;
     size_t pairs = 0;
-    uint16_t seen_keys = 0;
+    uint32_t seen_keys = 0;
     ZfCredentialDescriptorList allow_list = request->allow_list;
 
     memset(request, 0, sizeof(*request));
